equivalence exploration working. Truth tables confirm correct usage of laws. Can save and display graphs as svg. Basic coloring of nodes that provide 'new' edge options compared to a neighbor.
From here, I guess analysis of the graphs and further improvement of the graph display.

I could impliment a search for a specific stateement. Show how to get from A to B in the shortest path.

Could impliment an ML algo to guess how to find the equivalences without full exploration, but full exploration is so cheap and fast I am not motivated to do that.

I'd call this project semi-complete. I'm happy to do other stuff for now.